+++ kf_lib/fighting/fight/__init__.py
@@ -1,6 +1,7 @@
 from ._auto_fight import AutoFight  # used directly in testing scripts
 from ._base_fight import BaseFight  # for type annotations
-from ._free_for_all import AutoFreeForAll, free_for_all, NormalFreeForAll
+from ._free_for_all import AutoFreeForAll, AutoGroupFreeForAll, free_for_all, \
+    group_free_for_all, NormalFreeForAll, NormalGroupFreeForAll
 from ._helpers import fight, get_prefight_info, get_sides
 from ._sparring import spar
 from ._spectating import SpectateFight